backend/requirements.txt
# ── CareerPilot Backend Dependencies ──────────────────────────
# Install with:  pip install -r requirements.txt

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# ML model
sentence-transformers>=3.0.0
torch>=2.2.0

# Utilities (pulled by FastAPI, listed for clarity)
pydantic>=2.0.0

# LLM integration
google-generativeai>=0.8.0
python-dotenv>=1.0.0
